requirements.txt
fastapi
uvicorn
zhipuai
pinecone-client
python-dotenv